pi-gateway-two/requirements.txt
azure-iot-device>=2.12.0,<2.15.0
paho-mqtt>=1.6.1,<2.0.0
